# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PrepAI – AI Placement Mentor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
...
  /resume/analyze:
    post:
      operationId: analyzeResume
      tags: [resume]
      summary: Analyze a resume (text content)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeAnalysisInput"
      responses:
        "201":
          description: Resume analysis report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeReport"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /resume/reports:
    get:
      operationId: listResumeReports
      tags: [resume]
      summary: List resume analysis reports for the current user
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResumeReport"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /resume/reports/{id}:
    get:
      operationId: getResumeReport
      tags: [resume]
      summary: Get a specific resume report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resume report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeReport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # Mock Interviews
...
  /study-plans:
    get:
      operationId: listStudyPlans
      tags: [studyPlans]
      summary: List study plans for the current user
      responses:
        "200":
          description: List of study plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyPlan"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createStudyPlan
      tags: [studyPlans]
      summary: Generate a personalized study plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyPlanInput"
      responses:
        "201":
          description: Generated study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /study-plans/{id}:
    get:
      operationId: getStudyPlan
      tags: [studyPlans]
      summary: Get a study plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # Analytics
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get the current user's profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Dashboard
...
    ResumeReport:
      type: object
      required: [id, userId, fileName, atsScore, missingSkills, improvements, projectsFeedback, interviewQuestions, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        fileName:
          type: string
        atsScore:
          type: number
        missingSkills:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        projectsFeedback:
          type: string
        interviewQuestions:
          type: array
          items:
            type: string
        rawText:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ResumeAnalysisInput:
      type: object
      required: [fileName, resumeText]
      properties:
        fileName:
          type: string
          minLength: 1
        resumeText:
          type: string
          minLength: 1

    # Interviews
...
    StudyPlan:
      type: object
      required: [id, userId, targetCompany, daysLeft, weeks, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        targetCompany:
          type: string
        daysLeft:
          type: integer
        weeks:
          type: array
          items:
            $ref: "#/components/schemas/StudyWeek"
        createdAt:
          type: string
          format: date-time
    StudyWeek:
      type: object
      required: [week, topics, description]
      properties:
        week:
          type: integer
        topics:
          type: array
          items:
            type: string
        description:
          type: string
    StudyPlanInput:
      type: object
      required: [targetCompany, daysLeft]
      properties:
        targetCompany:
          type: string
          minLength: 1
        daysLeft:
          type: integer
          minimum: 1

    # Analytics
...
    UserProfile:
      type: object
      required: [userId, skills, targetCompany, targetRole, preparationProgress]
      properties:
        userId:
          type: string
        skills:
          type: array
          items:
            type: string
        projects:
          type: array
          items:
            type: string
        targetCompany:
          type: ["string", "null"]
        targetRole:
          type: ["string", "null"]
        preparationProgress:
          type: number
        bio:
          type: ["string", "null"]
        college:
          type: ["string", "null"]
        graduationYear:
          type: ["integer", "null"]
    UserProfileUpdate:
      type: object
      properties:
        skills:
          type: array
          items:
            type: string
        projects:
          type: array
          items:
            type: string
        targetCompany:
          type: string
        targetRole:
          type: string
        bio:
          type: string
        college:
          type: string
        graduationYear:
          type: integer