# lib/api-zod/src/generated/types/resumeAnalysisInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PrepAI – AI Placement Mentor API
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeAnalysisInput {
  /** @minLength 1 */
  fileName: string;
  /** @minLength 1 */
  resumeText: string;
}